Create README.md
The application has 2 screens. The first screen shows basic information related to prevention of COVID speed, if you want to get test for COVID you can click link, that leads to the page Invintro with all the information. And you can make an emergency call as well. The second screen shows the latest global and regional statistics with graphs.